// src/modules/promotions/enums/promotion.enum.ts
export enum PromotionStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SCHEDULED = 'SCHEDULED',
  EXPIRED = 'EXPIRED',
}

export enum PromotionConditionType {
  MIN_QTY = 'MIN_QTY',
  MIN_SPEND = 'MIN_SPEND',
  PRODUCT_COMBO = 'PRODUCT_COMBO',
  CATEGORY_COMBO = 'CATEGORY_COMBO',
  CUSTOMER_GROUP = 'CUSTOMER_GROUP',
  ALWAYS_TRUE = 'ALWAYS_TRUE',
}

export enum PromotionActionType {
  PERCENT_DISCOUNT = 'PERCENT_DISCOUNT',
  FIXED_DISCOUNT = 'FIXED_DISCOUNT',
  FIXED_PRICE = 'FIXED_PRICE',
  FREE_GIFT = 'FREE_GIFT',
  SET_TOTAL_PRICE = 'SET_TOTAL_PRICE',
}